+++ plone/recipe/codeanalysis/jscs.py
@@ -13,17 +13,7 @@ class JSCS(Analyser):
     @property
     def cmd(self):
         cmd = []
-
-        all_files = JSCS.split_lines(self.find_files('.*\.js'))
-        exc_files = []
-
-        # Should we exclude some files?
-        exclude = JSCS.split_lines(self.get_prefixed_option('exclude'))
-        if exclude:
-            exc_files = JSCS.split_lines(self.find_files('.*\.js', exclude))
-
-        # Remove excluded files
-        files = set(all_files) - set(exc_files)
+        files = self.find_files('.*\.js')
 
         if files:
             cmd.append(self.get_prefixed_option('bin'))
@@ -31,7 +21,7 @@ class JSCS(Analyser):
             if self.use_jenkins:
                 cmd.append('--reporter=checkstyle')
 
-            cmd.extend(list(files))
+            cmd.extend(files)
 
         return cmd