+++ plone/recipe/codeanalysis/tests/test_cleanlines.py
@@ -25,55 +25,50 @@ INVALID_TABS = """\
 class TestCleanLines(unittest.TestCase):
 
     def setUp(self):  # noqa
+        self.test_dir = mkdtemp()
         self.options = {
             'clean-lines': 'True',
             'clean-lines-exclude': '',
-            'jenkins': 'False'
+            'jenkins': 'False',
+            'directory': self.test_dir,
         }
-        self.test_dir = mkdtemp()
+        with open(os.path.join(self.test_dir, 'valid.py'), 'w') as f:
+            f.write(VALID_CODE)
 
     def tearDown(self):  # noqa
         rmtree(self.test_dir)
 
-    def _create_file_in_test_dir(self, filename, contents):
-        with open(os.path.join(self.test_dir, filename), 'w') as f:
-            f.write(contents)
-        self.options['directory'] = self.test_dir
-
     def test_analysis_should_return_false_if_trailing_spaces_are_found(self):
-        self._create_file_in_test_dir('invalid.py', INVALID_CODE)
+        with open(os.path.join(self.test_dir, 'invalid.py'), 'w') as f:
+            f.write(INVALID_CODE)
         self.assertFalse(CleanLines(self.options).run())
 
     def test_analysis_should_return_true_if_invalid_file_is_excluded(self):
         filename = 'invalid.py'
-        self._create_file_in_test_dir(filename, INVALID_CODE)
+        with open(os.path.join(self.test_dir, filename), 'w') as f:
+            f.write(INVALID_CODE)
         self.options['clean-lines-exclude'] = \
             '{0:s}/{1:s}'.format(self.test_dir, filename)
-
         self.assertTrue(CleanLines(self.options).run())
 
     def test_analysis_should_return_true_for_valid_files(self):
-        self._create_file_in_test_dir('valid.py', VALID_CODE)
         self.assertTrue(CleanLines(self.options).run())
 
     def test_analysis_should_return_false_if_file_with_tabs_is_found(self):
-        self._create_file_in_test_dir('invalid.xml', INVALID_TABS)
+        with open(os.path.join(self.test_dir, 'invalid.xml'), 'w') as f:
+            f.write(INVALID_TABS)
         self.assertFalse(CleanLines(self.options).run())
 
     def test_analysis_should_return_true_if_file_with_tabs_is_excluded(self):
         filename = 'invalid.xml'
-        self._create_file_in_test_dir(filename, INVALID_TABS)
+        with open(os.path.join(self.test_dir, filename), 'w') as f:
+            f.write(INVALID_TABS)
         self.options['clean-lines-exclude'] = \
             '{0:s}/{1:s}'.format(self.test_dir, filename)
-
         self.assertTrue(CleanLines(self.options).run())
 
     def test_analysis_should_return_true_for_non_watched_file_with_tabs(self):
         filename = 'invalid.abc'
-        self._create_file_in_test_dir(filename, INVALID_TABS)
-        self.assertTrue(CleanLines(self.options).run())
-
-    def test_analysis_should_return_true_for_non_watched_file_with_ts(self):
-        filename = 'invalid.abc'
-        self._create_file_in_test_dir(filename, INVALID_CODE)
+        with open(os.path.join(self.test_dir, filename), 'w') as f:
+            f.write(INVALID_TABS)
         self.assertTrue(CleanLines(self.options).run())