+++ CHANGES.rst
@@ -4,6 +4,9 @@ Change history
 2.0a3 (unreleased)
 ------------------
 
+- Exclude paths directly in ``find`` unix command which speeds up again a lot.
+  [saily]
+
 - Exclude empty strings in ``self.extensions`` which broke install with
   ``zc.buildout 1.7.1``. Fixes #115
   [saily]